requirements.txt
torch
torchvision
diffusers
transformers
accelerate
scipy
numpy
pillow
tqdm
scikit-image
lpips
